refactor: move image operations to dedicated files for maintainability
The implementations were untouched
